requirements.txt
asgiref==3.5.2
autopep8==2.0.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
da-vinci==0.3.0
Django==3.2.15
django-admin-thumbnails==0.2.6
django-ckeditor==6.3.2
django-daraja==1.2.0
django-js-asset==2.0.0
django-tinymce==3.5.0
idna==3.4
Pillow==9.2.0
pprintpp==0.4.0
pycodestyle==2.10.0
pycparser==2.21
PyMySQL==1.0.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
whitenoise==6.2.0
